reorganize utils; remove Cta component
