--- app/utils.js
@@ -1,90 +1,0 @@
-import dayjs from 'dayjs';
-
-export const noop = () => {};
-
-// string helpers
-export const makeKey = str => str.replace(/ /g, '-').toLowerCase();
-
-// date helpers
-export const getYearsAndMonths = (start, end) => {
-   const startDate = dayjs(start);
-   const endDate = end ? dayjs(end) : dayjs();
-   const years = endDate.diff(startDate, 'year');
-   const months = endDate.diff(startDate, 'month') % 12;
-   return { years, months };
-};
-export const getPrettyDate = date => dayjs(date).format('MMM YYYY');
-export const duration = (start, end, style) => {
-   const { years, months } = getYearsAndMonths(start, end);
-   const prettyStart = getPrettyDate(start);
-   const prettyEnd = end ? getPrettyDate(end) : 'Present';
-   const prettyDate = `${prettyStart} to ${prettyEnd}`;
-   const yoMo = `${years ? `${years} years ` : ''}${months} months`;
-   const approxYears = `~ ${Math.round(years + months / 12)} years`;
-
-   switch (style) {
-      case 'pretty':
-         return prettyDate;
-      case 'yoMo':
-         return yoMo;
-      case 'approxYears':
-         return approxYears;
-      default:
-         return `${prettyDate} (${yoMo})`;
-   }
-};
-export const getYearsSince = date => {
-   const now = dayjs();
-   const then = dayjs(date);
-   const years = now.diff(then, 'year');
-   // const months = now.diff(then, 'month') % 12;
-   return years;
-};
-
-// theme helpers
-export const colorCode = color => {
-   switch (color) {
-      case 'white':
-         return 'text-foreground';
-      case 'gray':
-         return 'text-bb-gray';
-      case 'blue':
-         return 'text-bb-blue';
-      case 'jade':
-         return 'text-bb-jade';
-      case 'jade-800':
-         return 'text-bb-jade-800';
-      case 'jade-700':
-         return 'text-bb-jade-700';
-      case 'jade-600':
-         return 'text-bb-jade-600';
-      case 'red':
-         return 'text-bb-red';
-      case 'teal':
-         return 'text-bb-teal';
-      case 'yellow':
-         return 'text-bb-yellow';
-      default:
-         return 'text-bb-jade';
-   }
-};
-
-export const maxWidthCode = size => {
-   switch (size) {
-      case 'wide':
-         return 'max-w-[1200px]';
-      case 'normal':
-         return 'max-w-[800px]';
-      case 'narrow':
-         return 'max-w-[640px]';
-      default:
-         return 'max-w-[800px]';
-   }
-};
-
-export const truncateString = (str, num) => {
-   if (str?.length <= num) {
-      return { newString: str, wasTruncated: false };
-   }
-   return { newString: str.slice(0, num) + '...', wasTruncated: true };
-};
