+++ components/common/index.ts
@@ -1,5 +1,4 @@
 export { default as Button } from './Button';
-export { default as Cta } from './Cta';
 export { default as PrimaryCta } from './PrimaryCta';
 export { default as CodeLink } from './CodeLink';
 export { default as Breadcrumb } from './Breadcrumb';
